feat: generated ebpfs go files

+++ pkg/ebpfmonitors/network_monitor.go
@@ -1,10 +1,10 @@
 package ebpfmonitors
 
-//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target bpf -go-package=ebpfmonitors NetworkMonitor bpf/network_monitor.bpf.c -- -I. -O2 -Wall -g
+//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target amd64,386,arm64,arm -go-package=gobpfs -output-dir=gobpfs NetworkMonitor bpf/network_monitor.bpf.c -- -Ibpf -O2 -Wall -g
 
 type NetworkEvent struct {
 	SrcIP    uint32
-	DestPort uint16
+	DestPort uint32
 }
 
 type networkMonitor struct {
